LayoutTests/js/global-resolve-through-eval-expected.txt

Make sure we do a correct property resolution of a global object property when contained by eval.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS accessGlobal() is true
PASS accessLocal() is true
PASS successfullyParsed is true

TEST COMPLETE
